Put everything in /animation to mimic master
