--- src/Components/.directory
@@ -1,5 +1,0 @@
-[Dolphin]
-HeaderColumnWidths=297,96,158
-Timestamp=2020,5,2,8,34,6
-Version=3
-ViewMode=1
