+++ src/Components/animation/DrillCuts.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import { StyleSheet, View } from 'react-native';
+import { StyleSheet, View, Animated, PanResponder } from 'react-native';
 import Svg, { Line, Circle } from 'react-native-svg';
-import { Animated, PanResponder } from 'react-native';
+
 import MovingCircle from './MovingCircle';
 
 /** The cuts that must be displayed at each step of a drill (a cut correspond to the position of a player at the previous step) */
@@ -29,7 +29,6 @@ class DrillCuts extends React.Component {
     this.cuts = [];
 
     if (this.props.drill.ids.length > 0) {
-
       var elemCount = this.props.drill.ids.length;
 
       /* For each step */
@@ -41,8 +40,7 @@ class DrillCuts extends React.Component {
         if (stepId > 0) {
           /* For each element displayed */
           for (var elemId = 0; elemId < elemCount; elemId++) {
-
-              console.log("drill cut, elem id: " + elemId);
+            console.log('drill cut, elem id: ' + elemId);
             /* If the element moves at this step */
             if (
               this.props.drill.positions[stepId][elemId] !== null &&
@@ -70,83 +68,80 @@ class DrillCuts extends React.Component {
 
                 elemCut.push(pos);
               }
-                
-                var counterCutX = (elemCut[0][0] + elemCut[1][0]) / 2;
-                var counterCutY = (elemCut[0][1] + elemCut[1][1]) / 2;
 
-                if(elemCut.length > 2){
-                    counterCutX = elemCut[2][0];
-                    counterCutY = elemCut[2][1];
-                }
+              var counterCutX = (elemCut[0][0] + elemCut[1][0]) / 2;
+              var counterCutY = (elemCut[0][1] + elemCut[1][1]) / 2;
 
-                
-                var x0 = elemCut[0][0];
-                var y0 = elemCut[0][1];
-                var x1 = elemCut[1][0];
-                var y1 = elemCut[1][1];
-                var x2 = counterCutX;
-                var y2 = counterCutY;
-                  
-                var d1 = Math.sqrt(Math.pow(x0 - x2, 2) + Math.pow(y0 - y2, 2));
-                var d2 = Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2));
+              if (elemCut.length > 2) {
+                counterCutX = elemCut[2][0];
+                counterCutY = elemCut[2][1];
+              }
 
-                var angle1 = (-Math.asin(Math.abs(y0 - y2) / d1)).toString() + "rad";
-                var angle2 = (-Math.asin(Math.abs(y1 - y2) / d2)).toString() + "rad";
+              var x0 = elemCut[0][0];
+              var y0 = elemCut[0][1];
+              var x1 = elemCut[1][0];
+              var y1 = elemCut[1][1];
+              var x2 = counterCutX;
+              var y2 = counterCutY;
 
-                var left1 = (x0 + x2 - d1) / 2;
-                var top1 = (y0 + y2) / 2;
-                var left2 = (x1 + x2 - d2) / 2;
-                var top2 = (y1 + y2) / 2;
+              var d1 = Math.sqrt(Math.pow(x0 - x2, 2) + Math.pow(y0 - y2, 2));
+              var d2 = Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2));
 
-                console.log("p0: " + x0 + "/" + y0);
-                console.log("p1: " + x1 + "/" + y1);
-                console.log("p2: " + x2 + "/" + y2);
-                console.log("angle1: " + angle1);
-                  
+              var angle1 = (-Math.asin(Math.abs(y0 - y2) / d1)).toString() + 'rad';
+              var angle2 = (-Math.asin(Math.abs(y1 - y2) / d2)).toString() + 'rad';
 
-                console.log("recompute cut");
+              var left1 = (x0 + x2 - d1) / 2;
+              var top1 = (y0 + y2) / 2;
+              var left2 = (x1 + x2 - d2) / 2;
+              var top2 = (y1 + y2) / 2;
 
-                
-                this.cuts[stepId].push({
-                    key: elemId,
-                    x0: x0,
-                    y0: y0,
-                    x1: x1,
-                    y1: y1,
-                    x2: x2,
-                    y2: y2,
-                    
-                    d1: d1,
-                    d2: d2,
+              console.log('p0: ' + x0 + '/' + y0);
+              console.log('p1: ' + x1 + '/' + y1);
+              console.log('p2: ' + x2 + '/' + y2);
+              console.log('angle1: ' + angle1);
 
-                    angle1: angle1,
-                    angle2: angle2,
+              console.log('recompute cut');
 
-                    left1: left1,
-                    top1: top1,
-                    left2: left2,
-                    top2: top2,
-                    
-                    cutCircle: new MovingCircle({
-                        onMoveEnd:this.props.onMoveEnd,
-                        elemId: elemId,
-                        animationHeight: this.props.animationHeight,
-                        animationWidth: this.props.animationWidth,
-                        cx: elemCut[1][0],
-                        cy: elemCut[1][1],
-                        radius: this.discRadius / 2,
-                        isCounterCut: false
-                    }),
-                    countercutCircle: new MovingCircle({
-                        onMoveEnd:this.props.onMoveEnd,
-                        elemId: elemId,
-                        animationHeight: this.props.animationHeight,
-                        animationWidth: this.props.animationWidth,
-                        cx: counterCutX,
-                        cy: counterCutY,
-                        radius: this.discRadius / 2,
-                        isCounterCut: true
-                    }),
+              this.cuts[stepId].push({
+                key: elemId,
+                x0,
+                y0,
+                x1,
+                y1,
+                x2,
+                y2,
+
+                d1,
+                d2,
+
+                angle1,
+                angle2,
+
+                left1,
+                top1,
+                left2,
+                top2,
+
+                cutCircle: new MovingCircle({
+                  onMoveEnd: this.props.onMoveEnd,
+                  elemId,
+                  animationHeight: this.props.animationHeight,
+                  animationWidth: this.props.animationWidth,
+                  cx: elemCut[1][0],
+                  cy: elemCut[1][1],
+                  radius: this.discRadius / 2,
+                  isCounterCut: false,
+                }),
+                countercutCircle: new MovingCircle({
+                  onMoveEnd: this.props.onMoveEnd,
+                  elemId,
+                  animationHeight: this.props.animationHeight,
+                  animationWidth: this.props.animationWidth,
+                  cx: counterCutX,
+                  cy: counterCutY,
+                  radius: this.discRadius / 2,
+                  isCounterCut: true,
+                }),
               });
             }
           }
@@ -180,13 +175,12 @@ class DrillCuts extends React.Component {
     }
   }
 
-    _display(element){
-        return (element.render());
-    }
-    
-    _displayCut = cut => {
+  _display(element) {
+    return element.render();
+  }
 
-        /*
+  _displayCut = cut => {
+    /*
 // In svg
 
 stroke="green" 
@@ -199,51 +193,53 @@ stroke="green"
               {this._display(cut.countercutCircle)}
             {this._display(cut.cutCircle)}
 */
-        
+
     /* If there is no counter cut */
-/*        console.log('pos 0: ' + cut.x0 + '/' + cut.y0);
+    /*        console.log('pos 0: ' + cut.x0 + '/' + cut.y0);
         console.log('pos 1: ' + cut.x1 + '/' + cut.y1);
         console.log('pos 2: ' + cut.x2 + '/' + cut.y2);
         console.log('disc radius: ' + this.discRadius);
 */
 
-        console.log("angle1: " + cut.angle1);
-        return (
-                <View key={cut.key+4000} style={[StyleSheet.absoluteFill]} height="100%" width="100%">
+    console.log('angle1: ' + cut.angle1);
+    return (
+      <View key={cut.key + 4000} style={[StyleSheet.absoluteFill]} height="100%" width="100%">
+        <View
+          style={[
+            { height: 1 },
+            { width: cut.d1 },
+            { borderRadius: 1 },
+            { borderWidth: 1 },
+            { borderColor: 'green' },
+            { borderStyle: 'dotted' },
+            { position: 'absolute' },
+            { top: cut.top1 },
+            { left: cut.left1 },
+            { transform: [{ rotate: cut.angle1 }] },
+          ]}
+        />
 
-                <View style={[
-                    {height: 1},
-                    {width: cut.d1},
-                    {borderRadius: 1},
-                    {borderWidth: 1},
-                    {borderColor: 'green'},
-                    {borderStyle: 'dotted'},
-                    {position: 'absolute'},
-                    {top:cut.top1},
-                    {left: cut.left1},
-                    {transform: [{ rotate: cut.angle1 }]}
-                ]} />
-                
-                <View style={[
-                    {height: 1},
-                    {width: cut.d2},
-                    {borderRadius: 1},
-                    {borderWidth: 1},
-                    {borderColor: 'green'},
-                    {borderStyle: 'dotted'},
-                    {position: 'absolute'},
-                    {top:cut.top2},
-                    {left: cut.left2},
-                    {transform: [{ rotate: cut.angle2 }]}
-                ]} />
-                
-            </View>
-        );
+        <View
+          style={[
+            { height: 1 },
+            { width: cut.d2 },
+            { borderRadius: 1 },
+            { borderWidth: 1 },
+            { borderColor: 'green' },
+            { borderStyle: 'dotted' },
+            { position: 'absolute' },
+            { top: cut.top2 },
+            { left: cut.left2 },
+            { transform: [{ rotate: cut.angle2 }] },
+          ]}
+        />
+      </View>
+    );
   };
 
   render() {
     return (
-        <View key="1" style={[{ position: 'absolute', left: 0, top: 0 }]} height="100%" width="100%">
+      <View key="1" style={[{ position: 'absolute', left: 0, top: 0 }]} height="100%" width="100%">
         {this.cuts.length >= this.props.currentStep &&
         this.cuts[this.props.currentStep] !== undefined &&
         this.cuts[this.props.currentStep] !== null
@@ -252,7 +248,7 @@ stroke="green"
       </View>
     );
   }
-    
+
   log() {
     for (var stepId = 0; stepId < this.cuts.length; stepId++) {
       console.log('step ' + stepId);
