+++ src/Components/animation/ElementAdderBar.js
